refactor(tui): split TuiRoot component into state + handler modules (<1000)
tui_component/mod.rs was a single 2594-line iocraft component. Bundle its
~104 shared `State` handles into a `Copy` `Ui` struct (state.rs) and move the
two giant event closures into their own modules:

- events.rs: gateway-event handling (apply_gw_event).
- keyboard.rs: terminal-event dispatch + dialog-focused key handling.
- keyboard_normal.rs: normal-mode (no dialog) key handling.

Each handler destructures `Ui` back into the original names so its body is
unchanged. mod.rs (575 lines) keeps the hook/state declarations, builds the
`Ui`, wires the two closures to the handlers, and renders the element tree.
No behavior change; all tui lib tests pass.
