+++ crates/rustyclaw-tui/src/app/tui_component/state.rs
@@ -1,0 +1,121 @@
+//! Shared state bundle for the TUI root component.
+//!
+//! [`Ui`] groups every `State<…>` handle declared in `TuiRoot` so the gateway-
+//! event and keyboard handlers can be split into their own modules. iocraft
+//! `State<T>` is `Copy`, so `Ui` is `Copy` and cheap to pass around. Handlers
+//! destructure it back into the original names, keeping their bodies verbatim.
+
+use std::collections::HashMap;
+use std::time::Instant;
+
+use iocraft::prelude::*;
+
+use crate::types::DisplayMessage;
+
+#[derive(Clone, Copy)]
+pub(super) struct Ui {
+    pub messages: State<Vec<DisplayMessage>>,
+    pub input_value: State<String>,
+    pub input_cursor_offset: State<usize>,
+    pub gw_status: State<rustyclaw_core::types::GatewayStatus>,
+    pub streaming: State<bool>,
+    pub stream_start: State<Option<Instant>>,
+    pub elapsed: State<String>,
+    pub scroll_offset: State<i32>,
+    pub spinner_tick: State<usize>,
+    pub should_quit: State<bool>,
+    pub streaming_buf: State<String>,
+    pub dynamic_model_label: State<Option<String>>,
+    pub dynamic_provider_id: State<Option<String>>,
+    pub selected_message_idx: State<Option<usize>>,
+    pub show_auth_dialog: State<bool>,
+    pub auth_code: State<String>,
+    pub auth_error: State<String>,
+    pub show_tool_approval: State<bool>,
+    pub tool_approval_id: State<String>,
+    pub tool_approval_name: State<String>,
+    pub tool_approval_args: State<String>,
+    pub tool_approval_selected: State<bool>,
+    pub show_vault_unlock: State<bool>,
+    pub vault_password: State<String>,
+    pub vault_error: State<String>,
+    pub hatching_dialog: State<rustyclaw_view::HatchingDialogData>,
+    pub show_pairing: State<bool>,
+    pub pairing_step: State<rustyclaw_view::PairingStep>,
+    pub pairing_field: State<rustyclaw_view::PairingField>,
+    pub pairing_public_key: State<String>,
+    pub pairing_fingerprint: State<String>,
+    pub pairing_fingerprint_art: State<String>,
+    pub pairing_qr_ascii: State<String>,
+    pub pairing_host: State<String>,
+    pub pairing_port: State<String>,
+    pub pairing_error: State<String>,
+    pub show_user_prompt: State<bool>,
+    pub user_prompt_id: State<String>,
+    pub user_prompt_title: State<String>,
+    pub user_prompt_desc: State<String>,
+    pub user_prompt_input: State<String>,
+    pub user_prompt_type: State<Option<rustyclaw_core::user_prompt_types::PromptType>>,
+    pub user_prompt_selected: State<usize>,
+    pub show_credential_request: State<bool>,
+    pub credential_request_id: State<String>,
+    pub credential_request_provider: State<String>,
+    pub credential_request_secret_name: State<String>,
+    pub credential_request_message: State<String>,
+    pub credential_request_input: State<String>,
+    pub show_provider_selector: State<bool>,
+    pub provider_selector_items: State<Vec<String>>,
+    pub provider_selector_ids: State<Vec<String>>,
+    pub provider_selector_hints: State<Vec<String>>,
+    pub provider_selector_cursor: State<usize>,
+    pub show_api_key_dialog: State<bool>,
+    pub api_key_provider: State<String>,
+    pub api_key_provider_display: State<String>,
+    pub api_key_input: State<String>,
+    pub api_key_help_url: State<String>,
+    pub api_key_help_text: State<String>,
+    pub show_device_flow: State<bool>,
+    pub device_flow_provider: State<String>,
+    pub device_flow_url: State<String>,
+    pub device_flow_code: State<String>,
+    pub device_flow_tick: State<usize>,
+    pub device_flow_browser_opened: State<bool>,
+    pub show_model_selector: State<bool>,
+    pub model_selector_provider: State<String>,
+    pub model_selector_provider_display: State<String>,
+    pub model_selector_models: State<Vec<String>>,
+    pub model_selector_cursor: State<usize>,
+    pub model_selector_loading: State<bool>,
+    pub threads: State<Vec<crate::action::ThreadInfo>>,
+    pub tab_focused: State<bool>,
+    pub tab_selected: State<usize>,
+    pub thread_messages_cache: State<HashMap<u64, Vec<DisplayMessage>>>,
+    pub foreground_thread_id: State<Option<u64>>,
+    pub command_completions: State<Vec<String>>,
+    pub command_selected: State<Option<usize>>,
+    pub model_completion_provider: State<Option<String>>,
+    pub model_completion_models: State<Vec<String>>,
+    pub model_completion_loading: State<Option<String>>,
+    pub prompt_attachments: State<Vec<rustyclaw_view::PromptAttachment>>,
+    pub show_secrets_dialog: State<bool>,
+    pub secrets_dialog_data: State<Vec<rustyclaw_view::SecretInfoData>>,
+    pub secrets_agent_access: State<bool>,
+    pub secrets_has_totp: State<bool>,
+    pub secrets_selected: State<Option<usize>>,
+    pub secrets_scroll_offset: State<usize>,
+    pub secrets_add_step: State<u8>,
+    pub secrets_add_name: State<String>,
+    pub secrets_add_value: State<String>,
+    pub show_skills_dialog: State<bool>,
+    pub skills_dialog_data: State<Vec<rustyclaw_view::SkillInfoData>>,
+    pub skills_selected: State<Option<usize>>,
+    pub show_details_dialog: State<bool>,
+    pub details_dialog_text: State<String>,
+    pub details_dialog_is_error: State<bool>,
+    pub details_dialog_scroll: State<usize>,
+    pub show_tool_perms_dialog: State<bool>,
+    pub tool_perms_dialog_data: State<Vec<rustyclaw_view::ToolPermInfoData>>,
+    pub tool_perms_selected: State<Option<usize>>,
+    pub skills_scroll_offset: State<usize>,
+    pub tool_perms_scroll_offset: State<usize>,
+}
